PictureToLatex/requirements.txt
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
transformers==4.50.3
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
accelerate==0.27.2
pix2text>=1.1.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.0.0
requests>=2.25.1 